refactor: move game dto package under game api
