+++ src/main/java/com/codenames/codenames/backend/game/api/GameSocketController.java
@@ -2,10 +2,10 @@ package com.codenames.codenames.backend.game.api;
 
 import com.codenames.codenames.backend.game.domain.Clue;
 import com.codenames.codenames.backend.game.application.GameService;
-import com.codenames.codenames.backend.game.dto.ClueMessage;
-import com.codenames.codenames.backend.game.dto.PassTurnMessage;
-import com.codenames.codenames.backend.game.dto.RevealCardMessage;
-import com.codenames.codenames.backend.game.dto.StartGameMessage;
+import com.codenames.codenames.backend.game.api.dto.ClueMessage;
+import com.codenames.codenames.backend.game.api.dto.PassTurnMessage;
+import com.codenames.codenames.backend.game.api.dto.RevealCardMessage;
+import com.codenames.codenames.backend.game.api.dto.StartGameMessage;
 import com.codenames.codenames.backend.recovery.application.SystemStatePersistenceService;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
