+++ src/main/java/com/codenames/codenames/backend/game/domain/GameManager.java
@@ -3,8 +3,8 @@ package com.codenames.codenames.backend.game.domain;
 import com.codenames.codenames.backend.game.domain.Clue;
 import com.codenames.codenames.backend.game.application.ClueValidationService;
 import com.codenames.codenames.backend.game.application.CardGenerator;
-import com.codenames.codenames.backend.game.dto.CardDataTransferObject;
-import com.codenames.codenames.backend.game.dto.GameStateDataTransferObject;
+import com.codenames.codenames.backend.game.api.dto.CardDataTransferObject;
+import com.codenames.codenames.backend.game.api.dto.GameStateDataTransferObject;
 import com.codenames.codenames.backend.game.domain.Color;
 import com.codenames.codenames.backend.lobby.domain.Role;
 import com.codenames.codenames.backend.lobby.domain.Team;
