+++ src/test/java/com/codenames/codenames/backend/game/api/GameSocketControllerTest.java
@@ -7,12 +7,12 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.codenames.codenames.backend.game.application.GameService;
-import com.codenames.codenames.backend.game.dto.ClueMessage;
-import com.codenames.codenames.backend.game.dto.PassTurnMessage;
-import com.codenames.codenames.backend.game.dto.RevealCardMessage;
-import com.codenames.codenames.backend.game.dto.StartGameMessage;
+import com.codenames.codenames.backend.game.api.dto.ClueMessage;
+import com.codenames.codenames.backend.game.api.dto.PassTurnMessage;
+import com.codenames.codenames.backend.game.api.dto.RevealCardMessage;
+import com.codenames.codenames.backend.game.api.dto.StartGameMessage;
 import com.codenames.codenames.backend.recovery.application.SystemStatePersistenceService;
-import com.codenames.codenames.backend.game.dto.GameStateDataTransferObject;
+import com.codenames.codenames.backend.game.api.dto.GameStateDataTransferObject;
 import com.codenames.codenames.backend.lobby.domain.Team;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
