+++ src/test/java/com/codenames/codenames/backend/game/application/GameServiceTest.java
@@ -12,10 +12,10 @@ import static org.mockito.Mockito.when;
 import com.codenames.codenames.backend.game.domain.Clue;
 import com.codenames.codenames.backend.game.domain.GameManager;
 import com.codenames.codenames.backend.game.domain.GameManagerFactory;
-import com.codenames.codenames.backend.game.dto.ClueDto;
-import com.codenames.codenames.backend.game.dto.CardDataTransferObject;
+import com.codenames.codenames.backend.game.api.dto.ClueDto;
+import com.codenames.codenames.backend.game.api.dto.CardDataTransferObject;
 import com.codenames.codenames.backend.game.mapping.DataTransferObjectService;
-import com.codenames.codenames.backend.game.dto.GameStateDataTransferObject;
+import com.codenames.codenames.backend.game.api.dto.GameStateDataTransferObject;
 import com.codenames.codenames.backend.game.domain.Color;
 import com.codenames.codenames.backend.lobby.domain.Role;
 import com.codenames.codenames.backend.lobby.domain.Team;
