+++ src/test/java/com/codenames/codenames/backend/game/domain/GameManagerTest.java
@@ -15,9 +15,9 @@ import static org.mockito.Mockito.when;
 import com.codenames.codenames.backend.game.domain.Clue;
 import com.codenames.codenames.backend.game.application.ClueValidationService;
 import com.codenames.codenames.backend.game.application.CardGenerator;
-import com.codenames.codenames.backend.game.dto.ClueDto;
-import com.codenames.codenames.backend.game.dto.CardDataTransferObject;
-import com.codenames.codenames.backend.game.dto.GameStateDataTransferObject;
+import com.codenames.codenames.backend.game.api.dto.ClueDto;
+import com.codenames.codenames.backend.game.api.dto.CardDataTransferObject;
+import com.codenames.codenames.backend.game.api.dto.GameStateDataTransferObject;
 import com.codenames.codenames.backend.game.domain.Color;
 import com.codenames.codenames.backend.lobby.domain.Role;
 import com.codenames.codenames.backend.lobby.domain.Team;
